Fix unit tests / cleanup
