+++ package.json
@@ -148,16 +148,21 @@
     "afterSign": "scripts/notarize.js"
   },
   "devDependencies": {
+    "@babel/core": "^7.23.0",
+    "@babel/plugin-transform-modules-commonjs": "^7.23.0",
     "@electron/notarize": "^2.1.0",
     "@next/bundle-analyzer": "^13.4.19",
     "@sentry/integrations": "^7.68.0",
     "@svgr/webpack": "^8.1.0",
-    "@testing-library/dom": "^9.3.1",
+    "@testing-library/dom": "^9.3.3",
     "@testing-library/jest-dom": "^6.1.3",
     "@testing-library/react": "^14.0.0",
     "@zeit/next-source-maps": "^0.0.3",
+    "babel-jest": "^29.7.0",
+    "babel-plugin-inline-react-svg": "^2.0.2",
+    "babel-plugin-styled-components": "^2.1.4",
     "cross-env": "^7.0.3",
-    "csv-parse": "^5.5.0",
+    "csv-parse": "^5.5.2",
     "debug": "^4.3.4",
     "dotenv": "^16.3.1",
     "electron": "^26.1.0",
@@ -171,14 +176,17 @@
     "hasha": "^5.2.2",
     "humanize": "^0.0.9",
     "icon-gen": "^4.0.0",
-    "jest": "^27.5.1",
-    "jest-canvas-mock": "^2.3.1",
-    "jest-css-modules-transform": "^4.2.1",
-    "jest-styled-components": "^7.0.4",
+    "jest": "^29.7.0",
+    "jest-canvas-mock": "^2.5.2",
+    "jest-css-modules-transform": "^4.4.2",
+    "jest-environment-jsdom": "^29.7.0",
+    "jest-esm-transformer": "^1.0.0",
+    "jest-esm-transformer-2": "^1.0.0",
+    "jest-styled-components": "^7.2.0",
     "markdown-to-jsx": "^7.3.2",
     "moment-range": "^4.0.2",
     "next": "^13.4.19",
-    "ooni-components": "1.0.0-alpha.17",
+    "ooni-components": "0.6.0-alpha.1",
     "os-locale": "^6.0.2",
     "prop-types": "^15.8.1",
     "raven-js": "^3.27.2",
@@ -196,7 +204,7 @@
     "react-table": "^7.0.0-rc.15",
     "react-test-renderer": "^18.2.0",
     "spectron": "^19.0.0",
-    "styled-components": "^6.0.8",
+    "styled-components": "^6.1.0",
     "stylis": "^4.3.0",
     "stylis-plugin-rtl": "^2.1.1",
     "use-clipboard-copy": "^0.2.0"
