+++ renderer/components/dashboard/__tests__/UrlList.test.js
@@ -1,14 +1,11 @@
-/**
- * @jest-environment jsdom
- */
-
 import React from 'react'
 import { theme } from 'ooni-components'
 import { ThemeProvider } from 'styled-components'
 import { IntlProvider } from 'react-intl'
-import { screen, render, fireEvent, cleanup } from '@testing-library/react'
+import { screen, render, fireEvent } from '@testing-library/react'
 import English from '../../../../lang/en.json'
 import { ipcRenderer } from 'electron'
+import api from '../../../../main/windows/api'
 
 import UrlList from '../UrlList'
 
@@ -30,23 +27,19 @@ const renderComponent = (component, locale = 'en', messages = English) => {
   )
 }
 
-describe('Tests for UrlList.js', () => {
+Object.defineProperty(window, 'electron', { value: api })
 
+describe('Tests for UrlList.js', () => {
   const websiteList = ['https://www.twitter.com', 'https://www.facebook.com']
 
-  beforeEach(() => {
-    renderComponent(<UrlList incomingList={null} />)
-  })
-  afterEach(() => {
-    cleanup()
-  })
-
   test('Run button is disabled by default', async () => {
+    renderComponent(<UrlList incomingList={null} />)
     const runButton = screen.getByTestId('button-run-custom-test')
     expect(runButton).toBeDisabled()
   })
 
   test('Run button is disabled on entering incorrect URL', async () => {
+    renderComponent(<UrlList incomingList={null} />)
     const urlInputBox = screen.getByRole('textbox')
     expect(urlInputBox).toBeInTheDocument()
     fireEvent.change(urlInputBox, {
@@ -57,6 +50,7 @@ describe('Tests for UrlList.js', () => {
   })
 
   test('Run button is enabled on entering correct URL', async () => {
+    renderComponent(<UrlList incomingList={null} />)
     const urlInputBox = screen.getByRole('textbox')
     expect(urlInputBox).toBeInTheDocument()
     fireEvent.change(urlInputBox, {
@@ -67,6 +61,7 @@ describe('Tests for UrlList.js', () => {
   })
 
   test('URL is correctly added and a new textbox is created', async () => {
+    renderComponent(<UrlList incomingList={null} />)
     const urlInputBox = screen.getByRole('textbox')
     expect(urlInputBox).toBeInTheDocument()
     fireEvent.change(urlInputBox, {
@@ -86,6 +81,7 @@ describe('Tests for UrlList.js', () => {
   })
 
   test('Added URL can be removed', async () => {
+    renderComponent(<UrlList incomingList={null} />)
     const urlInputBox = screen.getByRole('textbox')
     fireEvent.change(urlInputBox, {
       target: { value: websiteList[0] },
@@ -108,6 +104,7 @@ describe('Tests for UrlList.js', () => {
   })
 
   test('Custom websites test is triggered on entering a valid URL', async () => {
+    renderComponent(<UrlList incomingList={null} />)
     const urlInputBox = screen.getByRole('textbox')
     fireEvent.change(urlInputBox, {
       target: { value: websiteList[0] },
