+++ renderer/pages/_error.js
@@ -1,4 +1,3 @@
-/* global require */
 import React from 'react'
 import Error from 'next/error'
 import * as Sentry from '@sentry/electron/renderer'
