+++ scripts/__tests__/translations.test.js
@@ -1,11 +1,11 @@
 import glob from 'glob'
 import { basename } from 'path'
-import csvParse from 'csv-parse/lib/sync'
 import { readFileSync } from 'fs'
+import { parse } from 'csv-parse/sync'
 
 describe('Tests for translations', () => {
 
-  const langIdArray = csvParse(readFileSync('./data/lang-en.csv'), {from: 2}).map(row => row[0])
+  const langIdArray = parse(readFileSync('./data/lang-en.csv'), {from: 2}).map(row => row[0])
 
   const supportedLanguages = glob.sync('./lang/*.json').map((f) => basename(f, '.json'))
