+++ scripts/download-probe-cli.js
@@ -1,5 +1,3 @@
-/* global require */
-
 const path = require('path')
 const { execSync } = require('child_process')
 const { existsSync, ensureDirSync } = require('fs-extra')
