+++ scripts/download-tor.js
@@ -1,5 +1,3 @@
-/* global require */
-
 const path = require('path')
 const process = require('process')
 const { execSync } = require('child_process')
