+++ test/lib/setupTests.js
@@ -1,4 +1,5 @@
 import 'jest-canvas-mock'
 import 'jest-styled-components'
-import '@testing-library/jest-dom/extend-expect'
+import '@testing-library/jest-dom'
+
 jest.setTimeout(30000)
